[ZSH] Move back to rbenv from chruby
I found that rbenv is simpler to use than chruby, and more reliable when
changing folders, and its direct usage of ruby-build makes it simpler to
install Ruby versions.

Note: it's totally possible that I'm not using chruby right... But I
have no time to understand this, and rbenv does the job well enough for
my purposes.