[RUBY] Update gems and remove Fastlane
Close #1.

Removing Fastlane because if one needs it in a project, they
should fetch it locally to that project 😉.

The reason it was there in the first place was to install Xcode
plugins, but that's not something we can do anymore. unfortunately
Xcode 9 (I think) removed support for plugins. Developers should
now build extensions, but those are no where near as powerful. ☹️